pyproject.toml
[project]
name = "home-ops-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "fastapi>=0.128.0",
    "llama-index>=0.14.12",
    "llama-index-embeddings-openai>=0.5.1",
    "openai>=2.14.0",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.12.0",
    "ragas>=0.4.2",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.13"
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "integration: tests that require external resources (index, API)",
]
